% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{BrowserRouter}
\alias{BrowserRouter}
\title{BrowserRouter}
\usage{
BrowserRouter(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A BrowserRouter component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.BrowserRouter.html}
}
