% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{createBrowserRouter}
\alias{createBrowserRouter}
\title{createBrowserRouter}
\usage{
createBrowserRouter(...)
}
\arguments{
\item{...}{\code{\link{Route}} elements, typically wrapped in
\code{\link{createRoutesFromElements}}.}
}
\value{
A createBrowserRouter component.
}
\description{
\url{https://reactrouter.com/6.30.0/routers/create-browser-router}
}
\details{
Creates a browser router using the data router API.
Use with \code{\link{createRoutesFromElements}} and \code{\link{Route}}.
}
