% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{createHashRouter}
\alias{createHashRouter}
\title{createHashRouter}
\usage{
createHashRouter(...)
}
\arguments{
\item{...}{\code{\link{Route}} elements, typically wrapped in
\code{\link{createRoutesFromElements}}.}
}
\value{
A createHashRouter component.
}
\description{
\url{https://reactrouter.com/6.30.0/routers/create-hash-router}
}
\details{
Creates a hash router using the data router API.
Use with \code{\link{createRoutesFromElements}} and \code{\link{Route}}.
}
