% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Form}
\alias{Form}
\title{Form}
\usage{
Form(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A Form component.
}
\description{
\url{https://api.reactrouter.com/v7/variables/react-router.Form.html}
}
