% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{HashRouter}
\alias{HashRouter}
\title{HashRouter}
\usage{
HashRouter(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A HashRouter component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.HashRouter.html}
}
