% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{MemoryRouter}
\alias{MemoryRouter}
\title{MemoryRouter}
\usage{
MemoryRouter(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A MemoryRouter component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.MemoryRouter.html}
}
