% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Navigate}
\alias{Navigate}
\title{Navigate}
\usage{
Navigate(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A Navigate component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.Navigate.html}
}
