% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Outlet}
\alias{Outlet}
\title{Outlet}
\usage{
Outlet(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A Outlet component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.Outlet.html}
}
