% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{reactRouterDependency}
\alias{reactRouterDependency}
\title{react-router-dom JS dependency}
\usage{
reactRouterDependency()
}
\value{
HTML dependency object.
}
\description{
react-router-dom JS dependency
}
