% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Routes}
\alias{Routes}
\title{Routes}
\usage{
Routes(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A Routes component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.Routes.html}
}
