% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{createRoutesFromElements}
\alias{createRoutesFromElements}
\title{createRoutesFromElements}
\usage{
createRoutesFromElements(...)
}
\arguments{
\item{...}{\code{\link{Route}} elements.}
}
\value{
A tag list of Route elements.
}
\description{
\url{https://reactrouter.com/6.30.0/utils/create-routes-from-elements}
}
\details{
Wraps \code{\link{Route}} elements for use with \code{\link{createHashRouter}}.
The actual conversion from JSX elements to route objects happens on the
JavaScript side.
}
