% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{HashRouter}
\alias{HashRouter}
\title{HashRouter}
\usage{
HashRouter(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A HashRouter component.
}
\description{
\url{https://reactrouter.com/6.30.0/router-components/hash-router}
}
