% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/inputs.R
\name{Link}
\alias{Link}
\alias{Link.shinyInput}
\alias{updateLink.shinyInput}
\title{Link}
\usage{
Link(..., reloadDocument = TRUE)

Link.shinyInput(inputId, ..., reloadDocument = TRUE)

updateLink.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to element.}

\item{reloadDocument}{Boolean. Default TRUE. Let browser handle the transition normally}

\item{inputId}{ID of the component.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
A Link component.
}
\description{
The `reloadDocument` can be used to skip client side routing and let the 
browser handle the transition normally (as if it were an <a href>). In 
React Router v6 `reloadDocument` if `FALSE`, but given shiny behavior, the
`Link()` function makes it `TRUE` by default.

\url{https://reactrouter.com/6.30.0/components/link}
}
