% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/inputs.R
\name{Link}
\alias{Link}
\alias{Link.shinyInput}
\alias{updateLink.shinyInput}
\title{Link}
\usage{
Link(...)

Link.shinyInput(inputId, ...)

updateLink.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to element.}

\item{inputId}{ID of the component.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
A Link component.
}
\description{
\url{https://reactrouter.com/6.30.0/components/link}

\url{https://reactrouter.com/en/main/components/link}
}
