% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{MemoryRouter}
\alias{MemoryRouter}
\title{MemoryRouter}
\usage{
MemoryRouter(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A MemoryRouter component.
}
\description{
\url{https://reactrouter.com/6.30.0/router-components/memory-router}
}
