% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Navigate}
\alias{Navigate}
\title{Navigate}
\usage{
Navigate(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A Navigate component.
}
\description{
\url{https://reactrouter.com/6.30.0/components/navigate}
}
