% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/inputs.R
\name{NavLink}
\alias{NavLink}
\alias{NavLink.shinyInput}
\alias{updateNavLink.shinyInput}
\title{NavLink}
\usage{
NavLink(..., reloadDocument = FALSE)

NavLink.shinyInput(inputId, ..., reloadDocument = TRUE)

updateNavLink.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to element.}

\item{reloadDocument}{Boolean. Default TRUE. Let browser handle the transition normally}

\item{inputId}{ID of the component.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
A NavLink component.
}
\description{
The `reloadDocument` can be used to skip client side routing and let the 
browser handle the transition normally (as if it were an <a href>). Given 
shiny behavior, using `reloadDocument = TRUE` allows to render correctly
objects created in the server side.

\url{https://reactrouter.com/6.30.0/components/nav-link}
}
