% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/inputs.R
\name{NavLink}
\alias{NavLink}
\alias{NavLink.shinyInput}
\alias{updateNavLink.shinyInput}
\title{NavLink}
\usage{
NavLink(...)

NavLink.shinyInput(inputId, ...)

updateNavLink.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to element.}

\item{inputId}{ID of the component.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
A NavLink component.
}
\description{
\url{https://reactrouter.com/6.30.0/components/nav-link}

\url{https://reactrouter.com/en/main/components/nav-link}
}
