% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactRouterExample.R
\name{reactRouterExample}
\alias{reactRouterExample}
\title{Run reactRouterExample example}
\usage{
reactRouterExample(example = NULL, ...)
}
\arguments{
\item{example}{The name of the example to run, or `NULL` to retrieve the list of examples.}

\item{...}{Additional arguments to pass to `shiny::runApp()`.}
}
\value{
This function normally does not return;
interrupt R to stop the application (usually by pressing Ctrl+C or Esc).
}
\description{
Launch a Shiny example app or list the available examples.
Use `reactRouter::reactRouterExample("basic")` to run a showcase app.
}
\seealso{
[shiny.blueprint::runExample()] which this function is an adaptation.
}
