% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Route}
\alias{Route}
\title{Route}
\usage{
Route(..., element, loader = NULL, key = randomKey())
}
\arguments{
\item{...}{Props to pass to element.}

\item{element}{element wrapped in a `shiny::div()`.}

\item{loader}{Optional. A \code{\link{JS}} expression evaluating to a
loader function, e.g. \code{JS("({ params }) => fetch(...)")}. For a
plain unconditional redirect, use \code{\link{redirect}}. To embed
static R data, serialize it first with \code{jsonlite::toJSON()} and
wrap the result in \code{JS()}.}

\item{key}{By default uses a UUID key in the `div()` of the `element` arg.}
}
\value{
A Route component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.Route.html}
}
\details{
Internally the `element` is wrapped in a `shiny::div()`
with a UUID key so, in case R shiny is used, shiny can differentiate
each element.
}
