% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Route}
\alias{Route}
\title{Route}
\usage{
Route(..., element, loader = NULL, key = uuid::UUIDgenerate())
}
\arguments{
\item{...}{Props to pass to element.}

\item{element}{element wrapped in a `shiny::div()`.}

\item{loader}{Optional. Can be:
- A data.frame or list: serialized as a static JS loader
- An R function: called at build time (no params available)
- A JS() expression: passed through as-is}

\item{key}{By default uses a UUID key in the `div()` of the `element` arg.}
}
\value{
A Route component.
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.Route.html}
}
\details{
Internally the `element` is wrapped in a `shiny::div()`
with a UUID key so, in case R shiny is used, shiny can differentiate
each element.
}
