% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Route}
\alias{Route}
\title{Route}
\usage{
Route(..., element)
}
\arguments{
\item{...}{Props to pass to element.}

\item{element}{element with UUID key wrap in a `shiny::div()`.}
}
\value{
A Route component.
}
\description{
Internally the `element` is wrapped in a `shiny::div()` 
with a UUID key so, in case R shiny is used, shiny can differentiate 
each element.
}
