% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Route}
\alias{Route}
\title{Route}
\usage{
Route(...)
}
\arguments{
\item{...}{Props to pass to element.}
}
\value{
A Route component.
}
\description{
\url{https://reactrouter.com/6.30.0/components/route}
}
