% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{useHref}
\alias{useHref}
\title{useHref}
\usage{
useHref(into = NULL, as = "children", to, render = NULL, ...)
}
\arguments{
\item{into}{A component that will receive the value.}

\item{as}{Character. The name of the component's prop to inject into.}

\item{to}{Character. The path to resolve.}

\item{...}{Additional props to pass to the component.}
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.useHref.html}
}
\details{
Calls the \code{useHref()} hook and injects the resolved href string
\code{as} a prop of the \code{into} component.
}
