% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{useNavigationType}
\alias{useNavigationType}
\title{useNavigationType}
\usage{
useNavigationType(into = NULL, as = "children", render = NULL, ...)
}
\arguments{
\item{into}{A component (HTML tag or \pkg{shiny.react}-based element)
that will receive the hook value as the specified prop.}

\item{as}{Character. The name of the component's prop to inject the hook
value into. Defaults to \code{"children"}.}

\item{render}{Optional \code{\link{JS}} function \code{(value) => ReactNode}
used in place of \code{into}/\code{as}.}

\item{...}{Additional props to pass to the component.}
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.useNavigationType.html}
}
\details{
Calls the \code{useNavigationType()} hook and injects the result
\code{as} a prop of the \code{into} component.
Returns one of \code{"POP"}, \code{"PUSH"}, or \code{"REPLACE"}.
}
