% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useNavigationType}
\alias{useNavigationType}
\title{useNavigationType}
\usage{
useNavigationType(into, as = "children", ...)
}
\arguments{
\item{into}{A component that will receive the value.}

\item{as}{Character. The name of the component's prop to inject into.}

\item{...}{Additional props to pass to the component.}
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.useNavigationType.html}
}
\details{
Calls the \code{useNavigationType()} hook and injects the result
\code{as} a prop of the \code{into} component.
Returns one of \code{"POP"}, \code{"PUSH"}, or \code{"REPLACE"}.
}
