% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{useSearchParams}
\alias{useSearchParams}
\title{useSearchParams}
\usage{
useSearchParams(into = NULL, as = "children", param = NULL, render = NULL, ...)
}
\arguments{
\item{into}{A component that will receive the value.}

\item{as}{Character. The name of the component's prop to inject into.}

\item{param}{Character. Optional name of a single query parameter
to extract (e.g. \code{"q"}). If \code{NULL}, all params are
passed as a named list.}

\item{...}{Additional props to pass to the component.}
}
\description{
\url{https://api.reactrouter.com/v7/functions/react-router.useSearchParams.html}
}
\details{
Calls the \code{useSearchParams()} hook and injects the result
\code{as} a prop of the \code{into} component. Use the \code{param}
argument to extract a single query parameter by name.
}
